functions/ingest/requirements.txt
# Cloud Function dependencies
google-cloud-bigquery==3.12.0
